Move classes of the new bootstrapping to their own package
