+++ lib/src/main/java/org/sonarsource/scanner/lib/internal/facade/forked/package-info.java
@@ -17,34 +17,5 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.scanner.lib.internal;
-
-import com.google.gson.annotations.SerializedName;
-import javax.annotation.Nullable;
-
-public abstract class ResourceMetadata {
-  @SerializedName("filename")
-  private final String filename;
-  @SerializedName("sha256")
-  private final String sha256;
-  @SerializedName("downloadUrl")
-  private final String downloadUrl;
-
-  ResourceMetadata(String filename, String sha256, @Nullable String downloadUrl) {
-    this.filename = filename;
-    this.sha256 = sha256;
-    this.downloadUrl = downloadUrl;
-  }
-
-  public String getFilename() {
-    return filename;
-  }
-
-  public String getSha256() {
-    return sha256;
-  }
-
-  public String getDownloadUrl() {
-    return downloadUrl;
-  }
-}
+@javax.annotation.ParametersAreNonnullByDefault
+package org.sonarsource.scanner.lib.internal.facade.forked;
