+++ lib/src/test/java/org/sonarsource/scanner/lib/internal/facade/forked/JavaRunnerTest.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.scanner.lib.internal;
+package org.sonarsource.scanner.lib.internal.facade.forked;
 
 import java.nio.file.Paths;
 import java.util.List;
