+++ lib/src/test/resources/org/sonarsource/scanner/lib/internal/facade/forked/createRunner_jreProvisioning.json
